Create video_compilation_page.html
Build an app that is functionally similar to this example project. Try not to copy the example project, give it your own personal style.

Fulfill the user stories below and get all the tests to pass to complete the lab.

User Stories:

You should have a main element as the only child of the body element.
You should have an h1 element with the topic of your page.
You should have a paragraph introducing the topic of your page below your h1 element.
You should have three section elements below your first paragraph.
Each section should contain an h2 element, a paragraph, and an iframe element, in this order.
The three iframe elements should have a src attribute set to a valid video.
Each iframe element should also have a title attribute to describe the embedded content, and a height attribute and a width attribute to set the element to a proper size.