Create event_hub.html
Build an app that is functionally similar to this example project. Try not to copy the example project, give it your own personal style.

In this lab you will utilize the semantic HTML elements to create the structure of a web page. You'll add content and images to make it look like a real event hub.

Fulfil the user stories below and get all the tests to pass to complete the lab.

User Stories:

You should have a header element.

Inside the header element, you should have an h1 element that contains the text Event Hub, and a nav element.

Inside the nav element, you should have an unordered list of two items containing links to different sections of the page. The first item should have the text Upcoming Events, and the second item should have the text Past Events.

Each link should be represented by an a element with an href attribute that links to the corresponding section of the page, #upcoming-events and #past-events respectively.

You should have a main element that contains the different sections of the page.

Inside the main element, you should have two section elements.

The first section element should have an id attribute with the value upcoming-events

Inside the #upcoming-events section, you should have:

An h2 element with the text Upcoming Events.
Two article elements. Each article should represent an event, and it should have :
An h3 element for the event title.
A p element for the event description. You can add a date at the bottom if you like.
The second section element should have an id attribute with the value past-events.

Inside the #past-events section, you should have:

An h2 element with the text Past Events.
Two article elements. Each article element should represent a past event, and it should have:
An h3 element for the event title,
A p element for the event description. You can add a date at the bottom if you like.
An image element with the src attribute pointing to an image file and the alt attribute with a description of the image.

Note: You can use any text for the event descriptions and dates. You can use the following image URLs for the images if you like:

https://cdn.freecodecamp.org/curriculum/labs/past-event1.jpg.
https://cdn.freecodecamp.org/curriculum/labs/past-event2.jpg.