Create checkout_page.html
Build an app that is functionally similar to this example project. Try not to copy the example project, give it your own personal style.

Fulfill the user stories below and get all the tests to pass to complete the lab.

User Stories:

You should have an h1 element with the text Checkout.
You should have two section elements immediately after the h1 element.
You should have an h2 element with the text Your Cart within the first section.
You should have an image of an item in the first section with appropriate alternate text. You can use this image if you would like: https://cdn.freecodecamp.org/curriculum/labs/cube.jpg
You should have an h2 element with the text Payment Information within the second section.
You should have a form element within the second section.
You should have an input with an id and name of card-name within your form and a label associated with it.
You should have an input with an id and name of card-number within your form and a label associated with it.
At least two of your input elements should be required and have an aria-required attribute set to true.